requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pyserial
